requirements.txt
streamlit
pandas
statsmodels
openpyxl
numpy
scipy
plotly
python-dotenv
google-generativeai
markdown